refactoring of youtube cards
